+++ client/src/components/AboutDialog.tsx
@@ -9,9 +9,9 @@ type AboutDialogProps = {
 
 const VersionLine = (): JSX.Element => (
 	<p>
-		Version 0.2.1 (
+		Version 0.2.2 (
 		<a
-			href="https://github.com/serkonda7/teamotp/blob/main/CHANGELOG.md#021"
+			href="https://github.com/serkonda7/teamotp/blob/main/CHANGELOG.md#022"
 			target="_blank"
 			rel="noreferrer"
 		>